refactor(jsx): migrate to new type definition
