+++ src/jsx/components.test.tsx
@@ -5,7 +5,7 @@ import { HtmlEscapedCallbackPhase, resolveCallback as rawResolveCallback } from 
 import { ErrorBoundary } from './components'
 import { Suspense, renderToReadableStream } from './streaming'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { jsx } from './index'
+import { jsx } from '.'
 
 function resolveCallback(template: string | HtmlEscapedString) {
   return rawResolveCallback(template, HtmlEscapedCallbackPhase.Stream, false, {})
