+++ src/jsx/dom/components.ts
@@ -1,14 +1,16 @@
-import type { FC, Child } from '..'
+import type { FC, PropsWithChildren, Child } from '..'
 import type { FallbackRender, ErrorHandler } from '../components'
 import { DOM_ERROR_HANDLER } from '../constants'
 import { Fragment } from './jsx-runtime'
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
-export const ErrorBoundary: FC<{
-  fallback?: Child
-  fallbackRender?: FallbackRender
-  onError?: ErrorHandler
-}> = (({ children, fallback, fallbackRender, onError }: any) => {
+export const ErrorBoundary: FC<
+  PropsWithChildren<{
+    fallback?: Child
+    fallbackRender?: FallbackRender
+    onError?: ErrorHandler
+  }>
+> = (({ children, fallback, fallbackRender, onError }: any) => {
   const res = Fragment({ children })
   ;(res as any)[DOM_ERROR_HANDLER] = (err: any) => {
     if (err instanceof Promise) {
@@ -20,7 +22,10 @@ export const ErrorBoundary: FC<{
   return res
 }) as any
 
-export const Suspense: FC<{ fallback: any }> = (({ children, fallback }: any) => {
+export const Suspense: FC<PropsWithChildren<{ fallback: any }>> = (({
+  children,
+  fallback,
+}: any) => {
   const res = Fragment({ children })
   ;(res as any)[DOM_ERROR_HANDLER] = (err: any, retry: () => void) => {
     if (!(err instanceof Promise)) {
