+++ src/jsx/dom/context.test.tsx
@@ -1,10 +1,10 @@
 import { JSDOM } from 'jsdom'
+import { createContext as createContextCommon, useContext as useContextCommon } from '..' // for common
+import { use, Suspense } from '..'
 // run tests by old style jsx default
 // hono/jsx/jsx-runtime and hono/jsx/dom/jsx-runtime are tested in their respective settings
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { jsx, Fragment } from '..'
-import { createContext as createContextCommon, useContext as useContextCommon } from '..' // for common
-import { use, Suspense } from '..'
 import { createContext as createContextDom, useContext as useContextDom } from '.' // for dom
 import { render } from '.'
