+++ src/jsx/dom/css.test.tsx
@@ -3,10 +3,10 @@ import { JSDOM } from 'jsdom'
 // hono/jsx/jsx-runtime and hono/jsx/dom/jsx-runtime are tested in their respective settings
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { jsx } from '..'
+import type { JSXNode } from '..'
 import { Style, css, rawCssString, createCssContext } from '../../helper/css'
 import { minify } from '../../helper/css/common'
 import { renderTest } from '../../helper/css/common.test'
-import type { JSXNode } from '../../jsx'
 import { render } from '.'
 
 describe('Style and css for jsx/dom', () => {
