+++ src/jsx/dom/jsx-dev-runtime.ts
@@ -1,4 +1,4 @@
-import type { Props } from '..'
+import type { Props } from '../base'
 import { normalizeIntrinsicElementProps } from '../utils'
 
 export const jsxDEV = (tag: string | Function, props: Props, key: string | undefined) => {
